+++ src/Client/Crypto/CryptoClientException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Pastell\Client\Crypto;
-
-class CryptoClientException extends \Exception
-{
-}