+++ CHANGELOG.md
@@ -34,6 +34,7 @@ doivent être mises à jour #1667
 
 - Suppression de la table collectivite_fournisseur #2205
 - Suppression de validca et du code lié à l'authentification par certificat #1983
+- Suppression du code lié à libersign et la signature locale #1983
 
 # [4.1.11] - 2025-07-15